moved useGetIsPasswordVerified hook
