+++ src/views/LHD/components/LHDModal/index.tsx
@@ -10,7 +10,7 @@ import { icons } from 'components/uikit/Svg/types'
 // hooks and actions
 import useDebounce from 'hooks/useDebounce'
 import { useSetLhdAuth } from 'state/lhd/hooks/useGetIsLhdAuth'
-import useGetIsPasswordVerified from 'hooks/queries/useGetIsPasswordVerified'
+import useGetIsPasswordVerified from 'state/lhd/hooks/useGetIsPasswordVerified'
 
 const SOCIAL_LINKS: { icon: icons; href: string }[] = [
   { icon: icons.TWITTER, href: 'https://twitter.com/ape_swap' },
